test(api): fix concurrency
